cial_task/requirements.txt
Django>=4.1,<4.2
djangorestframework
psycopg2
requests
beautifulsoup4
gunicorn